requirements.txt
pillow
numpy
pytorch